image_selecter*: Add radio buttons for choosing red/blue channel launch/master_sync.launch: Add respawn launch/start.launch: Fusion between 3, 4 and 9                     Calculate image_width, image_height, and meter_per_pixel by pixel_scale include/potential_field_utils.hpp: Correct Sharr operator with respect to new transformation of image tf src/image_to_vectorfield_node.cpp: Better heuristic for repelling force src/image_viewer_node.cpp: Fix viz by adding rgb2bgr conversion src/odom_to_vectorfield_node.cpp: Casting of integer position src/vectorfield_merger_node.cpp: Add verbose ouput for misssized vector fields src/vectorfield_to_kinematic_node.cpp: Fix robot position with respect to new tf pattern: Add alternative images launch: Add launch scripts for 3,4,9; 3,4,6,7,9; and 3,4,6,7,9 with two images config/config.rviz: Update

+++ potential_field_navigation/src/odom_to_vectorfield_node.cpp
@@ -42,11 +42,10 @@ void process(const nav_msgs::OdometryConstPtr &odom) {
   cv::Point2i pose2d;
 //    pose2d = cv::Point2i((int) (odom->pose.pose.position.x / meterPerPixel + imageWidth/2), (int) (odom->pose.pose.position.y / meterPerPixel + imageHeight/2));
   if (pixelMode) {
-    pose2d = cv::Point2i(odom->pose.pose.position.x * pixelScale, odom->pose.pose.position.y * pixelScale);
+    pose2d = cv::Point2i((int) (odom->pose.pose.position.x * pixelScale), (int) (odom->pose.pose.position.y * pixelScale));
   } else {
     pose2d = cv::Point2i((int) (imageWidth / 2 - odom->pose.pose.position.y / meterPerPixel), (int) (imageHeight / 2 - odom->pose.pose.position.x / meterPerPixel));
   }
-
 
   if (firstRun) {
     firstRun = false;
